Re-write local preprocessing method.
This improves the preprocessing to correctly assign FreeSurfer regions of
interest to the `y` value and co-variates to the `X` value. This also sets up
support for multiple co-variates by employing a new `getNormalizedTags` function
and conditionally adding bias to `X`. Finally, it makes file reading
asynchronous and returns a Promise.